% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_strings.R
\name{correct_wrong_chars}
\alias{correct_wrong_chars}
\title{Correct wrong characters}
\usage{
correct_wrong_chars(string)
}
\arguments{
\item{string}{A vector character;}
}
\value{
A string corrected.
}
\description{
Correct some characters errors with str_replace_all.
}
\details{
Some strings are wrong in mercadoedu databases, so we replace this characters by correct characters.
}
\examples{
\dontrun{
  if(interactive()){
    dplyr::mutate(data,
                  "correct_name" = name |>
                                   correct_strings())
  }
}
}
\seealso{
\code{\link[stringr]{str_replace_all}}
}
