% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_cs}
\alias{get_cs}
\title{Get course table}
\usage{
get_cs(con, tbl_name = "course")
}
\arguments{
\item{con}{A database connection;}

\item{tbl_name}{A string of table name in database;\strong{\preformatted{Default: "course"
}

}}
}
\value{
The course table from database in data frame format.
}
\description{
Get course table from mercadoedu's database(AWS Redshift).
}
\details{
Sometimes you will have to reconnect to the database.
}
\examples{
\dontrun{
  if(interactive()){
    odbc::odbc() |>
      DBI::dbConnect("Amazon RDS Model",
                     timeout = 0) |>
      mercadoedu.tc::get_cs()
  }
}

}
\seealso{
\code{\link[dplyr]{tbl}},\code{\link[dplyr]{select}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{collect}}
}
