% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_csalias}
\alias{get_csalias}
\title{Get coursealias table}
\usage{
get_csalias(con, tbl_name = "coursealias")
}
\arguments{
\item{con}{A database connection;}

\item{tbl_name}{A string of table name in database;\strong{\preformatted{Default: "coursealias"
}

}}
}
\value{
The coursealias table from database in data frame format.
}
\description{
Get coursealias table from mercadoedu's database(AWS Redshift).
}
\details{
Sometimes you will have to reconnect to the database.
}
\examples{
\dontrun{
  if(interactive()){
    odbc::odbc() |>
      DBI::dbConnect("Amazon RDS Model",
                     timeout = 0) |>
      mercadoedu.tc::get_csalias()
  }
}

}
\seealso{
\code{\link[dplyr]{tbl}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{collect}}
}
