% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_ftk}
\alias{get_ftk}
\title{Get fromtokey table}
\usage{
get_ftk(con, tbl_name = "me_v3_pricing_fromtokey")
}
\arguments{
\item{con}{A database connection;}

\item{tbl_name}{A string of table name in database;\strong{\preformatted{Default: "me_v3_pricing_fromtokey"
}

}}
}
\value{
The fromtokey table from database in data frame format.
}
\description{
Get fromtokey table with alias_id(merged) from mercadoedu's database(AWS Redshift).
}
\details{
Sometimes you will have to reconnect to the database.
}
\examples{
\dontrun{
  if(interactive()){
    odbc::odbc() |>
      DBI::dbConnect("Amazon RDS Model",
                     timeout = 0) |>
      mercadoedu.tc::get_ftk()
  }
}

}
\seealso{
\code{\link[dplyr]{tbl}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{collect}},\code{\link[dplyr]{mutate-joins}}
}
