% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_pc}
\alias{get_pc}
\title{Get pricing_course table}
\usage{
get_pc(con, tbl_name = "pricing_course")
}
\arguments{
\item{con}{A database connection;}

\item{tbl_name}{A string of table name in database;\strong{\preformatted{Default: "pricing_course"
}

}}
}
\value{
The pricing_course table from database in data frame format.
}
\description{
Get pricing_course table from mercadoedu's database(AWS Redshift).
}
\details{
Sometimes you will have to reconnect to the database.
}
\examples{
\dontrun{
  if(interactive()){
    odbc::odbc() |>
      DBI::dbConnect("Amazon RDS Model",
                     timeout = 0) |>
      mercadoedu.tc::get_pc()
  }
}

}
\seealso{
\code{\link[dplyr]{tbl}},\code{\link[dplyr]{select}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{collect}}
}
