% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Get stopwords table}
\usage{
get_stopwords(con, tbl_name = "model_stopwords")
}
\arguments{
\item{con}{A database connection;}

\item{tbl_name}{A string of table name in database;\strong{\preformatted{Default: "model_stopwords"
}

}}
}
\value{
The stopwords table from database in data frame format.
}
\description{
Get stopwords table from mercadoedu's database(AWS RDS).
}
\details{
Sometimes you will have to reconnect to the database.
}
\examples{
\dontrun{
  if(interactive()){
    odbc::odbc() |>
      DBI::dbConnect("Amazon RDS Model",
                     timeout = 0) |>
      mercadoedu.tc::get_stopwords()
  }
}

}
\seealso{
\code{\link[dplyr]{tbl}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{collect}}
}
