% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_databases.R
\name{mount_frankenstein}
\alias{mount_frankenstein}
\title{Mount frankenstein table}
\usage{
mount_frankenstein(con, stopwords_con, checks = FALSE)
}
\arguments{
\item{con}{A database connection;}

\item{stopwords_con}{A database connection of stopwords table;}

\item{checks}{A logical that defines when to check tables;\strong{\preformatted{Default: FALSE
}

}}
}
\value{
A frankenstein table from mercadoedu's database.
}
\description{
Join and append multiples table from mercadoedu's database.
}
\details{
This function adds the same stopwords inputted but without accent.
}
\examples{
\dontrun{
  if(interactive()){
    frankenstein <- odbc::odbc() |>
      DBI::dbConnect("Amazon RDS mercadoedu",
                     timeout = 0) |>
        mercadoedu.tc::mount_frankenstein(stopwords_con = odbc::odbc() |>
                                            DBI::dbConnect("Amazon RDS Model",
                                                           timeout = 0))
 }
}

}
\seealso{
\code{\link[dplyr]{filter-joins}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{reexports}}
\code{\link[stringr]{str_replace}},\code{\link[stringr]{case}},\code{\link[stringr]{str_trim}}
\code{\link[tm]{removeWords}}
}
